+++ e2e/tests/desktop/registered-shopper.spec.js
@@ -49,17 +49,31 @@ test("Registered shopper can checkout items", async ({ page }) => {
   await page.goto(config.RETAIL_APP_HOME);
 
   await page.getByRole("link", { name: "Womens" }).hover();
+  const topsNav = await page.getByRole("link", { name: "Tops", exact: true });
+  await expect(topsNav).toBeVisible();
 
-  await page.getByRole("link", { name: "Tops" }).click();
+  await topsNav.click();
 
-  await expect(page.getByRole("heading", { name: "Tops" })).toBeVisible();
+  // PLP
+  const cableKnitShell = await page.getByRole("link", {
+    name: /Cable Knit Shell/i,
+  });
+  // selecting swatch
+  const initialImgEl = await cableKnitShell.locator("img");
+  const initialSrc = await initialImgEl.getAttribute("src");
+  await expect(cableKnitShell.getByText(/From £44\.16/i)).toBeVisible();
 
-  await page.getByRole("link", { name: /Stripe Shell/i }).click();
+  await cableKnitShell.getByLabel(/Ivory/, { exact: true }).hover();
+  const changedImgEl = await cableKnitShell.locator("img");
+  const changeImgSrc = await changedImgEl.getAttribute("src");
+  await expect(cableKnitShell.getByText(/From £44\.16/i)).toBeVisible();
+  expect(changeImgSrc).not.toBe(initialSrc);
+  await cableKnitShell.click();
 
+  // PDP
   await expect(
-    page.getByRole("heading", { name: /Stripe Shell/i })
+    page.getByRole("heading", { name: /Cable Knit Shell/i })
   ).toBeVisible();
-
   await page.getByRole("radio", { name: "L", exact: true }).click();
 
   await page.locator("button[data-testid='quantity-increment']").click();
@@ -69,7 +83,7 @@ test("Registered shopper can checkout items", async ({ page }) => {
   const updatedPageURL = await page.url();
   const params = updatedPageURL.split("?")[1];
   expect(params).toMatch(/size=9LG/i);
-  expect(params).toMatch(/color=JJ5YPA7/i);
+  expect(params).toMatch(/color=JJ5AAXX/i);
 
   await page.getByRole("button", { name: /Add to Cart/i }).click();
 
@@ -81,7 +95,9 @@ test("Registered shopper can checkout items", async ({ page }) => {
 
   await page.getByLabel(/My cart/i).click();
 
-  await expect(page.getByRole("link", { name: /Stripe Shell/i })).toBeVisible();
+  await expect(
+    page.getByRole("link", { name: /Cable Knit Shell/i })
+  ).toBeVisible();
 
   await page.getByRole("link", { name: "Proceed to Checkout" }).click();
 
@@ -172,5 +188,7 @@ test("Registered shopper can checkout items", async ({ page }) => {
     page.getByRole("heading", { name: /Order Summary/i })
   ).toBeVisible();
   await expect(page.getByText(/2 Items/i)).toBeVisible();
-  await expect(page.getByRole("link", { name: /Stripe Shell/i })).toBeVisible();
+  await expect(
+    page.getByRole("link", { name: /Cable Knit Shell/i })
+  ).toBeVisible();
 });