+++ packages/template-retail-react-app/app/pages/cart/index.jsx
@@ -67,7 +67,8 @@ const Cart = () => {
         {
             parameters: {
                 ids: productIds,
-                allImages: true
+                allImages: true,
+                perPricebook: true
             }
         },
         {